backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
numpy==1.26.4
Pillow==10.2.0
opencv-python-headless==4.8.0.76
deepface==0.0.88
scipy==1.11.3
soundfile==0.12.1
librosa==0.10.2
torch==2.2.2
transformers==4.42.0
whisper==20240307
gTTS==2.3.1
requests==2.31.0
